Cast to long for JRuby < 1.1.2 (vvs).
git-svn-id: svn://rubyforge.org/var/svn/mongrel/branches/stable_1-1@1017 19e92222-5c0b-0410-8929-a290d50e31e9

+++ ext/http11_java/org/jruby/mongrel/Http11.java
@@ -247,7 +247,7 @@ public class Http11 extends RubyObject {
             if(this.hp.has_error()) {
                 throw new RaiseException(runtime, eHttpParserError, "Invalid HTTP format, parsing fails.", true);
             } else {
-                return runtime.newFixnum(this.hp.parser.nread);
+                return runtime.newFixnum((long)this.hp.parser.nread);
             }
         }
     }
